[validation] remove mviewmempool backend for coins_to_uncache
At this point in validation, we want to remove the backend in order to
prevent going to CoinsTip() unnecessarily which could mean forgetting
to add coins to coins_to_uncache. However, it's not necessary to remove
m_view's backend (which is m_viewmempool), we should be removing
m_viewmempool's backend (which is CoinsTip()).